LayoutTests/http/tests/security/contentSecurityPolicy/script-src-none-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'allow'.

CONSOLE MESSAGE: Refused to load the script 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/script.js' because it violates the following Content Security Policy directive: "script-src 'none'".

Loads an iframe which in turns tries to load an external script. The iframe has a content security policy disabling external scripts. So the script should not get executed.



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS
